Gloabl Exception Handler, CustomBaseException,ProductNotFound&ProductNotValid

+++ src/main/java/com/example/demo/products/commandHandlers/DeleteProductCommandHandler.java
@@ -1,11 +1,14 @@
 package com.example.demo.products.commandHandlers;
 
 import com.example.demo.Command;
+import com.example.demo.Exceptions.ProductNotFoundException;
 import com.example.demo.products.Model.Product;
 import com.example.demo.products.ProductRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
+
+import java.util.Optional;
 
 @Service
 public class DeleteProductCommandHandler implements Command <Integer, ResponseEntity>{
@@ -15,6 +18,10 @@ public class DeleteProductCommandHandler implements Command <Integer, ResponseEn
 
     @Override
     public ResponseEntity<ResponseEntity> execute(Integer id) {
+        Optional<Product> optionalProduct = productRepository.findById(id);
+        if(optionalProduct.isEmpty()){
+            throw new ProductNotFoundException();
+        }
 
         Product product = productRepository.findById(id).get();
         productRepository.delete(product);
